pyproject.toml
[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[project]
name = "projektveckor-portal"
version = "0.1.0"
description = "Teacher-first project-week portal hosted on hule.education (FastAPI + Vue/Vite)."
authors = [{ name = "Olof Larsson", email = "paunchygent@gmail.com" }]
readme = "README.md"
requires-python = ">=3.13,<3.15"
license = { text = "MIT" }

dependencies = [
  "fastapi",
  "uvicorn[standard]",
  "pydantic",
  "pydantic-settings",
  "pyyaml",
  "markdown",
  "httpx",
  "dishka",
]

[dependency-groups]
dev = [
  "ruff",
  "mypy",
  "pytest",
  "pytest-cov",
  "types-PyYAML",
  "types-Markdown",
]

[tool.pdm.scripts]
dev = "uvicorn projektveckor_portal.app:app --host 127.0.0.1 --port 8000 --reload"
"dev:lan" = "uvicorn projektveckor_portal.app:app --host 0.0.0.0 --port 8000 --reload"
serve = "uvicorn projektveckor_portal.app:app --host 127.0.0.1 --port 8000"
"serve:lan" = "uvicorn projektveckor_portal.app:app --host 0.0.0.0 --port 8000"
lint = "ruff check ."
format = "ruff format ."
typecheck = "mypy ."
test = "pytest -q"

# Doc-as-code
validate-docs = "python -m scripts.docs_as_code.validate_docs"
validate-backlog = "python -m scripts.docs_as_code.validate_tasks"
validate-skills = "python -m scripts.docs_as_code.validate_skill_frontmatter"
index-backlog = "python -m scripts.docs_as_code.index_tasks --root docs/backlog --out docs/backlog/README-index.md --fail-on-missing"
new-doc = "python -m scripts.docs_as_code.new_doc"
new-task = "python -m scripts.docs_as_code.new_task"
new-programme = "python -m scripts.docs_as_code.new_task --type programme"
new-epic = "python -m scripts.docs_as_code.new_task --type epic"
new-story = "python -m scripts.docs_as_code.new_task --type story"
new-review = "python -m scripts.docs_as_code.new_task --type review"
new-fix = "python -m scripts.docs_as_code.new_task --type fix"
new-rule = "python -m scripts.docs_as_code.new_rule"

# Markdown quality (Docforge-style: prettier + markdownlint-cli2)
"lint:md" = "python -m scripts.maintenance.markdown_quality lint"
"fix:md" = "python -m scripts.maintenance.markdown_quality fix"
"format:md" = "python -m scripts.maintenance.markdown_quality format"
"check:md" = "python -m scripts.maintenance.markdown_quality check"
"autofix:md" = "python -m scripts.maintenance.markdown_quality autofix"

# Frontend (run from repo root; Skriptoteket-style)
"frontend:install" = "pnpm -C frontend install"
"frontend:dev" = "pnpm -C frontend dev"
"frontend:dev:lan" = "pnpm -C frontend dev:lan"
"frontend:build" = "pnpm -C frontend build"
"frontend:preview" = "pnpm -C frontend preview"
"frontend:preview:lan" = "pnpm -C frontend preview:lan"
"frontend:typecheck" = "pnpm -C frontend typecheck"

[tool.ruff]
line-length = 100
target-version = "py313"
exclude = [
  ".git",
  ".venv",
  "__pycache__",
  ".mypy_cache",
  ".pytest_cache",
  ".ruff_cache",
  "frontend/node_modules",
  "frontend/dist",
]

[tool.ruff.lint]
select = ["E", "W", "F", "I"]
ignore = ["E203"]

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
check_untyped_defs = true
explicit_package_bases = true
namespace_packages = true
mypy_path = "src"
exclude = ["frontend/"]

[tool.pytest.ini_options]
pythonpath = ["src"]
